Changed create-conversation to typescript
